Refactor the Request and Response class #82
