+++ api/helpers/Request.js
@@ -1,0 +1,9 @@
+class Request {
+  constructor(query = {}, params = {}, body = {}) {
+    this.query = query;
+    this.body = body;
+    this.params = params;
+  }
+}
+
+module.exports = Request;
